requirements.txt
sentence-transformers
fastapi
uvicorns